Basic auth support added
* refactor nodejs authorization lambda
+ added basic autentication code
+ param store setup to keep basic auth details
