+++ ife-cognito/variables.tf
@@ -35,6 +35,11 @@ variable "certificate_arn" {
   default     = ""
 }
 
+variable "param_store_client_prefix" {
+  description = "Prefix used in parameter store where all client basic auth configurations will be stored"
+  type        = string
+}
+
 variable "tags" {
   description = "A map of tags to add to all resources"
   type        = map(string)
