+++ ife-lambda/variables.tf
@@ -13,6 +13,11 @@ variable "env_user_pool_id" {
   type        = string
 }
 
+variable "param_store_client_prefix" {
+  description = "Prefix used in parameter store where all client basic auth configurations will be stored"
+  type        = string
+}
+
 variable "tags" {
   description = "A map of tags to add to all resources"
   type        = map(string)
